fixed typo, implemented ability to have different data of same name, implemented ability to optionally save used file

+++ devito/interfaces.py
@@ -478,17 +478,28 @@ class MemmapManager():
     """Class for managing all memmap related settings"""
     # used to enable memmap as default
     _use_memmap = False
+    # determine whether created files are deleted by default
+    _delete_file = True
     # flag for registering exit func
     _registered = False
     _default_disk_path = os.path.join(gettempdir(), "devito_disk")
     # contains str name of all memmap file created
-    _created_data = set()
+    _created_data = {}
+    # unique id
+    _id = 0
     _default_exit_code = 0
 
     @staticmethod
     def set_memmap(memmap):
         """Call this method to set default value of memmap"""
         MemmapManager._use_memmap = memmap
+
+    @staticmethod
+    def set_delete_file(delete):
+        """Call this method to set default flag contolling whether to delete
+        crated files.
+        """
+        MemmapManager._delete_file = delete
 
     @staticmethod
     def set_default_disk_path(default_disk_path):
@@ -499,11 +510,12 @@ class MemmapManager():
     def setup(data_self, *args, **kwargs):
         """This method is used to setup memmap parameters for data classes.
 
-        :param name: Name of data, must be unique
+        :param name: Name of data
         :param memmap: Boolean indicates whether memmap is used. Optional
-        :param disk_path: String indicates path to create memmap file. Optional
+        :param disk_path: String indicates directory to create memmap file. Optional
+        :param delete_file: Boolean indicates whether to delete created file. Optional
 
-        Note: If memmap or disk_path are not provided, the default values
+        Note: If memmap, disk_path or delete_file are not provided, the default values
         are used.
         """
         data_self.memmap = kwargs.get('memmap', MemmapManager._use_memmap)
@@ -514,22 +526,25 @@ class MemmapManager():
             if not os.path.exists(disk_path):
                 os.makedirs(disk_path)
 
-            data_self.f = disk_path + "/data_" + kwargs.get('name')
-            MemmapManager._created_data.add(data_self.f)
+            data_self.f = disk_path + "/data_" + kwargs.get('name') + "_" + str(MemmapManager._id)
+            MemmapManager._id += 1
+            data_self.delete_file = kwargs.get('delete_file', MemmapManager._delete_file)
+            MemmapManager._created_data[data_self.f] = data_self.delete_file
 
             if not MemmapManager._registered:
                 MemmapManager._register_remove_memmap_file_signal()
                 MemmapManager._registered = True
 
     @staticmethod
-    def _reomve_memmap_file():
+    def _remove_memmap_file():
         """This method is used to clean up memmap file"""
         for f in MemmapManager._created_data:
-            try:
-                os.remove(f)
-            except OSError:
-                print("error removing " + f + " it may be already removed, skipping")
-                pass
+            if MemmapManager._created_data[f]:
+                try:
+                    os.remove(f)
+                except OSError:
+                    print("error removing " + f + " it may be already removed, skipping")
+                    pass
 
     @staticmethod
     def _remove_memmap_file_on_signal(*args):
@@ -539,7 +554,7 @@ class MemmapManager():
     @staticmethod
     def _register_remove_memmap_file_signal():
         """This method is used to register clean up method for chosen signals"""
-        atexit.register(MemmapManager._reomve_memmap_file)
+        atexit.register(MemmapManager._remove_memmap_file)
 
         for sig in (SIGABRT, SIGINT, SIGSEGV, SIGTERM):
             signal(sig, MemmapManager._remove_memmap_file_on_signal)
